# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cinema Booking System API",
        "description": "API documentation for Cinema Booking System",
        "contact": {
            "email": "support@cinemat.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{protocol}://{host}",
            "description": "API Server",
            "variables": {
                "protocol": {
                    "enum": [
                        "http",
                        "https"
                    ],
                    "default": "http"
                },
                "host": {
                    "default": "localhost:8089"
                }
            }
        }
    ],
    "paths": {
...
        "/api/promotions/validate": {
            "post": {
                "tags": [
                    "Promotions"
                ],
                "summary": "Xác thực và áp dụng mã giảm giá",
                "operationId": "77d02cd7641fff6c5875bfad0f6ea2e1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "booking_id"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "SUMMER2024"
                                    },
                                    "booking_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mã giảm giá hợp lệ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "promotion": {
                                                    "type": "object"
                                                },
                                                "discount_amount": {
                                                    "type": "number"
                                                },
                                                "total_amount": {
                                                    "type": "number"
                                                },
                                                "final_amount": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Mã giảm giá không hợp lệ"
                    },
                    "403": {
                        "description": "Không có quyền"
                    },
                    "404": {
                        "description": "Mã giảm giá không tồn tại"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}